Refacter dashboard components
With this change, we are refactoring dashboard.metadata component as  dashboard.core and dashboard.metadata.api component as dashboard.api
