+++ components/org.wso2.carbon.dashboards.core/pom.xml
@@ -1,0 +1,112 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+ ~ Copyright 2016 (c) WSO2 Inc. (http://wso2.org) All Rights Reserved.
+ ~
+ ~ Licensed under the Apache License, Version 2.0 (the "License");
+ ~ you may not use this file except in compliance with the License.
+ ~ You may obtain a copy of the License at
+ ~
+ ~      http://www.apache.org/licenses/LICENSE-2.0
+ ~
+ ~ Unless required by applicable law or agreed to in writing, software
+ ~ distributed under the License is distributed on an "AS IS" BASIS,
+ ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ ~ See the License for the specific language governing permissions and
+ ~ limitations under the License.
+-->
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <groupId>org.wso2.carbon.dashboards</groupId>
+        <artifactId>carbon-dashboards</artifactId>
+        <version>3.0.0-SNAPSHOT</version>
+        <relativePath>../../pom.xml</relativePath>
+    </parent>
+
+    <modelVersion>4.0.0</modelVersion>
+    <groupId>org.wso2.carbon.dashboards</groupId>
+    <artifactId>org.wso2.carbon.dashboards.core</artifactId>
+    <packaging>bundle</packaging>
+    <name>WSO2 Carbon - Dashboards metadata store</name>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.wso2.carbon.uuf</groupId>
+            <artifactId>org.wso2.carbon.uuf.core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.wso2.msf4j</groupId>
+            <artifactId>msf4j-core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.testng</groupId>
+            <artifactId>testng</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.messaging</groupId>
+            <artifactId>org.wso2.carbon.messaging</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-log4j12</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.datasources</groupId>
+            <artifactId>org.wso2.carbon.datasource.core</artifactId>
+        </dependency>
+    </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.felix</groupId>
+                <artifactId>maven-bundle-plugin</artifactId>
+                <extensions>true</extensions>
+                <configuration>
+                    <instructions>
+                        <Import-Package>
+                            org.wso2.carbon.uuf.*;version="${carbon.uuf.core.version}",
+                            org.osgi.framework.*;version="[1.8.0, 2.0.0)",
+                            <!--org.wso2.carbon.messaging;version="${carbon.messaging.version}",-->
+                            <!--TODO: replace below line with above commented line after carbon-messaging fix export version issue.-->
+                            org.wso2.carbon.messaging;version="1.0.5",
+                            org.wso2.carbon.datasource.core.*;version="${org.wso2.carbon.datasource.version.range}",
+                            org.wso2.msf4j.*,
+                            javax.ws.rs.*,
+                            io.netty.handler.*,
+                            org.osgi.service.component.annotations.*,
+                            org.yaml.snakeyaml,
+                            com.google.common.*;version="${guava.version.range}",
+                            org.slf4j.*;version="${slf4j.version.range}",
+                            javax.sql.*
+                        </Import-Package>
+                        <Export-Package>
+                            org.wso2.carbon.dashboards.core.*;version="${carbon.dashboards.version}"
+                        </Export-Package>
+                        <Carbon-Component>
+                            osgi.service;objectClass="org.wso2.carbon.dashboards.core.provider.MetadataProvider"
+                        </Carbon-Component>
+                    </instructions>
+                </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.6.0</version>
+                <configuration>
+                  <source>1.8</source>
+                  <target>1.8</target>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+</project>
