+++ components/org.wso2.carbon.dashboards.core/src/main/java/org/wso2/carbon/dashboards/core/internal/dao/utils/DAOUtils.java
@@ -1,0 +1,152 @@
+/*
+ *  Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ *
+ */
+package org.wso2.carbon.dashboards.core.internal.dao.utils;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.wso2.carbon.dashboards.core.exception.MetadataException;
+import org.wso2.carbon.datasource.core.api.DataSourceService;
+import org.wso2.carbon.datasource.core.exception.DataSourceException;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+/**
+ * Util class to define DB related common operation
+ */
+public class DAOUtils {
+
+    private static final DAOUtils instance = new DAOUtils();
+
+    private static final Logger log = LoggerFactory.getLogger(DAOUtils.class);
+
+    private static final String WSO2_DASHBOARD_DB = "WSO2_DASHBOARD_DB";
+
+    private DataSourceService dataSourceService;
+    private DataSource dataSource;
+
+
+    public DataSource getDataSource(String dataSourceName) throws DataSourceException {
+
+        if (dataSourceService == null) {
+            throw new RuntimeException("Datasource service is null. Cannot retrieve data source");
+        }
+        return (DataSource) dataSourceService.getDataSource(dataSourceName);
+    }
+
+    public static void setDataSourceService(DataSourceService dataSourceService) {
+        instance.dataSourceService = dataSourceService;
+    }
+
+
+    private DAOUtils() {
+        super();
+    }
+
+    public static DAOUtils getInstance() throws MetadataException {
+        instance.initialize(WSO2_DASHBOARD_DB);
+        return instance;
+    }
+
+    public void initialize(String dataSourceName) throws MetadataException {
+        if (dataSource != null) {
+            return;
+        }
+
+        try {
+            dataSource = instance.getDataSource(dataSourceName);
+        } catch (DataSourceException e) {
+            throw new MetadataException("Error in initializing database !", e);
+        }
+
+    }
+
+    public Connection getConnection() throws SQLException {
+        if (dataSource != null) {
+            return dataSource.getConnection();
+        }
+        throw new SQLException("Data source is not configured properly.");
+    }
+
+    /**
+     * Utility method to close the connection streams.
+     *
+     * @param preparedStatement PreparedStatement
+     * @param connection        Connection
+     * @param resultSet         ResultSet
+     */
+    public static void closeAllConnections(PreparedStatement preparedStatement, Connection connection,
+                                           ResultSet resultSet) {
+        closeConnection(connection);
+        closeResultSet(resultSet);
+        closeStatement(preparedStatement);
+    }
+
+    /**
+     * Close Connection
+     *
+     * @param dbConnection Connection
+     */
+    private static void closeConnection(Connection dbConnection) {
+        if (dbConnection != null) {
+            try {
+                dbConnection.close();
+            } catch (SQLException e) {
+                log.warn("Database error. Could not close database connection. Continuing with " +
+                        "others. - " + e.getMessage(), e);
+            }
+        }
+    }
+
+    /**
+     * Close ResultSet
+     *
+     * @param resultSet ResultSet
+     */
+    private static void closeResultSet(ResultSet resultSet) {
+        if (resultSet != null) {
+            try {
+                resultSet.close();
+            } catch (SQLException e) {
+                log.warn("Database error. Could not close ResultSet  - " + e.getMessage(), e);
+            }
+        }
+
+    }
+
+    /**
+     * Close PreparedStatement
+     *
+     * @param preparedStatement PreparedStatement
+     */
+    public static void closeStatement(PreparedStatement preparedStatement) {
+        if (preparedStatement != null) {
+            try {
+                preparedStatement.close();
+            } catch (SQLException e) {
+                log.warn("Database error. Could not close PreparedStatement. Continuing with" +
+                        " others. - " + e.getMessage(), e);
+            }
+        }
+
+    }
+}
