+++ components/org.wso2.carbon.dashboards.core/src/main/resources/sql/mysql.sql
