+++ components/org.wso2.carbon.dashboards.core/src/test/java/org/wso2/carbon/dashboards/core/internal/provider/impl/MetadataProviderImplTest.java
@@ -1,0 +1,76 @@
+/*
+ *  Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ *
+ */
+package org.wso2.carbon.dashboards.core.internal.provider.impl;
+
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Test;
+
+public class MetadataProviderImplTest {
+
+    @BeforeMethod
+    public void setUp() throws Exception {
+
+    }
+
+    @AfterMethod
+    public void tearDown() throws Exception {
+
+    }
+
+    @Test
+    public void testIsExists() throws Exception {
+
+    }
+
+    @Test
+    public void testUpdate() throws Exception {
+
+    }
+
+    @Test
+    public void testAdd() throws Exception {
+
+    }
+
+    @Test
+    public void testDelete() throws Exception {
+
+    }
+
+    @Test
+    public void testGet() throws Exception {
+
+    }
+
+    @Test
+    public void testGet1() throws Exception {
+
+    }
+
+    @Test
+    public void testRegisterDataSourceService() throws Exception {
+
+    }
+
+    @Test
+    public void testUnregisterDataSourceService() throws Exception {
+
+    }
+}
