+++ components/org.wso2.carbon.dashboards.view/src/main/fragments/dashboard-view/public/js/dashboard.js
@@ -148,7 +148,7 @@
     };
 
     /**
-     * generate the metadata payload to invoke rest apis
+     * generate the core payload to invoke rest apis
      * @returns metadata payload
      */
     var metaDataPayloadGeneration = function () {
