+++ features/org.wso2.carbon.dashboards.api.feature/src/main/resources/repository/database/DASHBOARD_DB.trace.db
@@ -1,0 +1,78 @@
+01-06 15:47:45 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-08 10:15:12 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-09 09:57:24 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 11:01:00 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 11:12:23 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 11:17:38 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 11:23:14 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 11:30:56 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 11:39:22 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 11:42:29 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 11:45:24 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 13:44:15 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 13:48:59 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 14:24:23 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 14:49:59 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 15:29:59 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 16:31:33 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 16:43:10 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 16:47:47 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-10 17:15:50 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-16 11:25:17 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-16 11:35:16 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-16 11:35:48 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-16 11:59:23 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-16 12:17:46 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
+01-16 13:37:50 jdbc[2]: SQLException
+org.h2.jdbc.JdbcSQLException: Table "DASHBOARD_RESOURCE" not found; SQL statement:
+ DROP TABLE DASHBOARD_RESOURCE [42102-140]
